# crawlers/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Linkedin parser API",
      "version": "0.0.1",
      "description": "Linkedin parser API"
    },
    "paths": {
...
      "/user/create": {
        "post": {
          "summary": "Create user.",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "body",
              "type": "object",
              "example": {
                  "login": "servicelogin@gsuit.com",
                  "password": "mypass1234"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "examples": {
                "application/json": "{}"
              }
            }
          }
        }
      },
...
      "/sn/search/scribe": {
        "post": {
          "summary": "Scribe from Linkedin Sales Navigator search EACH profile.",
          "parameters": [
            {
                "name": "Authorization",
                "in": "header",
                "description": "Auth token",
                "required": true,
                "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "description": "body",
              "required": true,
              "type": "object",
              "example": {
                  "credentials_id": "123",
                  "input_data": {
                      "campaign_data": {
                          "search_url": "www.linkedin.com/user/search=marketers",
                          "interval_pages": 10
                      }
                  }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "examples": {
                "application/json": "{}"
              }
            }
          }
        }
      }, 
...
    },
...
    "consumes": [
      "application/json"
    ]
  }  